# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trading Journal & Analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/dashboard:
    get:
      operationId: getDashboardAnalytics
      tags: [analytics]
      summary: Get dashboard summary metrics
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMetrics"
  /analytics/equity-curve:
    get:
      operationId: getEquityCurve
      tags: [analytics]
      summary: Get equity curve data
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Equity curve data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquityCurveResponse"
...
  /analytics/calendar:
    get:
      operationId: getCalendarData
      tags: [analytics]
      summary: Get calendar view data
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Calendar data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarResponse"
...
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: Get all goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalRequest"
      responses:
        "200":
          description: Goal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Goal deleted
...
components:
  schemas:
...
    DashboardMetrics:
      type: object
      properties:
        netPnl:
          type: number
        grossPnl:
          type: number
        totalTrades:
          type: integer
        winningTrades:
          type: integer
        losingTrades:
          type: integer
        winRate:
          type: number
        profitFactor:
          type: number
        avgWin:
          type: number
        avgLoss:
          type: number
        expectancy:
          type: number
        maxDrawdown:
          type: number
        maxDrawdownPct:
          type: number
        avgHoldingMinutes:
          type: number
        currentStreak:
          type: integer
        currentStreakType:
          type: string
          enum: [WIN, LOSS, NONE]
        longestWinStreak:
          type: integer
        longestLossStreak:
          type: integer
        totalCommissions:
          type: number
        sharpeRatio:
          type: number
      required:
        - netPnl
        - grossPnl
        - totalTrades
        - winningTrades
        - losingTrades
        - winRate
        - profitFactor
        - avgWin
        - avgLoss
        - expectancy
        - maxDrawdown
        - maxDrawdownPct
        - avgHoldingMinutes
        - currentStreak
        - currentStreakType
        - longestWinStreak
        - longestLossStreak
        - totalCommissions
        - sharpeRatio
    EquityCurvePoint:
      type: object
      properties:
        date:
          type: string
        cumulativePnl:
          type: number
        drawdown:
          type: number
        drawdownPct:
          type: number
        dailyPnl:
          type: number
        tradeCount:
          type: integer
      required:
        - date
        - cumulativePnl
        - drawdown
        - drawdownPct
        - dailyPnl
        - tradeCount
    EquityCurveResponse:
      type: object
      properties:
        points:
          type: array
          items:
            $ref: "#/components/schemas/EquityCurvePoint"
        startingBalance:
          type: number
      required:
        - points
        - startingBalance
...
    CalendarDay:
      type: object
      properties:
        date:
          type: string
        netPnl:
          type: number
        tradeCount:
          type: integer
        isWinDay:
          type: boolean
      required:
        - date
        - netPnl
        - tradeCount
        - isWinDay
    CalendarResponse:
      type: object
      properties:
        days:
          type: array
          items:
            $ref: "#/components/schemas/CalendarDay"
        monthlyNetPnl:
          type: number
        tradingDays:
          type: integer
        winDays:
          type: integer
        lossDays:
          type: integer
      required:
        - days
        - monthlyNetPnl
        - tradingDays
        - winDays
        - lossDays
...
    Goal:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [PNL, WIN_RATE, PROFIT_FACTOR, TRADE_COUNT, MAX_LOSS]
        name:
          type: string
        targetValue:
          type: number
        currentValue:
          type: number
        period:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY]
        startDate:
          type: string
        endDate:
          type: string
        isCompleted:
          type: boolean
        progressPct:
          type: number
        createdAt:
          type: string
      required:
        - id
        - type
        - name
        - targetValue
        - currentValue
        - period
        - startDate
        - endDate
        - isCompleted
        - progressPct
        - createdAt
    CreateGoalRequest:
      type: object
      properties:
        type:
          type: string
          enum: [PNL, WIN_RATE, PROFIT_FACTOR, TRADE_COUNT, MAX_LOSS]
        name:
          type: string
        targetValue:
          type: number
        period:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY]
        startDate:
          type: string
        endDate:
          type: string
      required:
        - type
        - name
        - targetValue
        - period
        - startDate
        - endDate
    UpdateGoalRequest:
      type: object
      properties:
        targetValue:
          type: number
        name:
          type: string
...