# lib/api-zod/src/generated/types/calendarResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Journal & Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarDay } from "./calendarDay";

export interface CalendarResponse {
  days: CalendarDay[];
  monthlyNetPnl: number;
  tradingDays: number;
  winDays: number;
  lossDays: number;
}